Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> Do we want to implement retry logic within music? Or push that to the client?

I actually think atomicPut is a client of Music and not part of Music itself (there is no mention of atomicPut in the protocol in the paper).

In fact, retrying atomicPut itself is not a good idea in case of failure, because the acquireLock and criticalPut phases might have completed successfully and only the releaseLock phase should be retried.

Patch-set: 3
Reviewer: Gerrit User 4904 <4904@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0