Update patch set 1
Patch Set 1:

I am thinking we can just remove this Method, its never used. It was intended to be used in a different way. We only use the one method in here and all the others are deprecated. 
Since this class was taken from portal, we don't use everything the same way. 
Does everyone agree, there is no need for this method?

Patch-set: 1