Update patch set 2
Patch Set 2:

Can you please include the fix as a patch in this request so we don't have to merge 2 pull requests?
You should be able to amend the pull request to fix a previous push. If it's too difficult, please abandon both pull requests and create a new one with the correct code.
Also there are other similar errors in your new code, not just the one I pointed out.

Patch-set: 2