Update patch set 5
Patch Set 5:

The confusing part is that this whole class is duplicated in both music-rest and music-core. The functions that you've added here in the core are the same as what's already been done in music-rest. I'd prefer if we remove one of these classes, probably from the music-rest.

Do you think you could do that with this pull request as well?

Also the "for testing only" functions shouldn't be public. They should either be protected or no modifier type

Patch-set: 5
Reviewer: Gerrit User 4079 <4079@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>